sign in sign up

+++ PiDev/src/main/java/com/pi/dev/serviceInterface/IRoleService.java
@@ -6,12 +6,11 @@ import com.pi.dev.models.Role;
 
 public interface IRoleService {
 
-	
 	List<Role> findAll();
 
 	Role addRole(Role role);
 
-	Role updateRole(Role role, Long roleID);
+	Role updateRole(Role role, Integer roleID);
 
 	void deleteRoleById(Long roleID);
 }
